+++ packages/experimental/src/components/ExampleComponent/ExampleComponent.js
@@ -5,13 +5,14 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { Button } from 'carbon-components-react';
+import { Button, ButtonSet } from 'carbon-components-react';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import { expPrefix } from '../../global/js/settings';
+import { pkgPrefix } from '../../global/js/settings';
+import cx from 'classnames';
 
-const blockClass = `${expPrefix}-example-component`;
+const blockClass = `${pkgPrefix}-example-component`;
 
 // import './example-component.scss'; // Do not import SCSS directly it will be rolled up separately.
 
@@ -44,11 +45,11 @@ export const ExampleComponent = ({
   };
 
   return (
-    <div
-      className={[`${blockClass}`, `${blockClass}--${size}`, mode].join(' ')}
+    <ButtonSet
+      className={cx([blockClass, `${blockClass}--${size}`, mode])}
       style={{
         /* stylelint-disable-next-line carbon/theme-token-use */
-        [`--${expPrefix}-border-color`]: borderColor,
+        [`--${pkgPrefix}-border-color`]: borderColor,
       }}
       {...props}>
       <Button
@@ -65,7 +66,7 @@ export const ExampleComponent = ({
         disabled={props.disabled}>
         {primaryButtonLabel}
       </Button>
-    </div>
+    </ButtonSet>
   );
 };